requirements.txt
attrs==24.2.0
aws-cdk-lib==2.162.1
aws-cdk.asset-awscli-v1==2.2.208
aws-cdk.asset-kubectl-v20==2.1.3
aws-cdk.asset-node-proxy-agent-v6==2.1.0
aws-cdk.cloud-assembly-schema==38.0.1
boto3==1.35.44
botocore==1.35.44
cattrs==24.1.2
constructs==10.4.2
importlib_resources==6.4.5
jmespath==1.0.1
jsii==1.104.0
publication==0.0.3
python-dateutil==2.9.0.post0
s3transfer==0.10.3
six==1.16.0
typeguard==2.13.3
typing_extensions==4.12.2
urllib3==2.2.3
